Move methods into themed files and require them in template.rb
