+++ file_creator.rb
@@ -1,0 +1,5 @@
+$path = File.expand_path(File.dirname(__FILE__))
+
+def render_file(path)
+  file = IO.read(path)
+end
