+++ gemfile_configurator.rb
@@ -1,0 +1,11 @@
+def remove_gemfile
+  remove_file "Gemfile"
+end
+
+def api_gemfile
+  file "Gemfile", render_file("#{$path}/files/Gemfile_api_only")
+end
+
+def integrated_app_gemfile
+  file "Gemfile", render_file("#{$path}/files/Gemfile")
+end
