+++ initialize_git.rb
@@ -1,0 +1,3 @@
+def initialize_git
+  git :init
+end
